Refactor<settings>: Modularized settings manager and updated the ui
